Updated tvim(). Added invim().
tvim() now records the pane id in $TVIM.

If $TVIM already exists, that pane is selected.
If $TVIM doesn't exist, or the pane has been closed, a new pane
is created and assigned to $TVIM.

invim() calls tmux send-keys on the TVIM pane.
If TVIM doesn't already exist, tvim is called first.