Move vim/general after vim/$plugins
This means vim/general overrides the language plugins, which may not
be what we want.

The reason for this is I want to make : not a keyword in perl, so that
Dist::Zilla::Plugin::Test::Changes is not all one word.

see: 031a912189ff0438e8a36e2cbb087a96071e1e5c